Moving seeding script generation into seeding folder
