+++ multiset/seeding/template.py
